requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
jsonify==0.5
MarkupSafe==1.0
Pillow==5.4.1
pprint==0.1
Werkzeug==0.14.1
